DAO smart contract code
This is a basic implementation of a DAO (Decentralized Autonomous Organization) smart contract that allows members to submit proposals, vote on proposals, and execute proposals if they are approved. The contract also includes functions for contributing funds to the DAO and withdrawing funds from the DAO.

To use this contract, members can submit proposals by calling the submitProposal function with a proposal description. Once a proposal is submitted, members can vote on the proposal using the voteProposal function. If a proposal receives more votes in favor than against, it can be executed using the executeProposal function.

In addition to the proposal functions, this contract also includes functions for contributing funds to the DAO using the contribute function and withdrawing funds from the DAO using the withdrawFunds function.

This contract provides a starting point for building more complex DAOs with additional features. It is written in Solidity and can be deployed on a blockchain that supports smart contracts.